Added event creation/editing
And some state-ful things to make things turn off/on.

Things still not quite working essentialy lie with me figuring out
how to copy an array of objects (eg mentors) to a state-ful object
to then update as needed. Currently keep going in loops with either
re-rendering infinite blocks, or undefined. This be a Saturday mentor
question :D